+++ compatibility/Residence/src/main/java/me/Fupery/ArtMap/Compatibility/impl/ResidenceCompat.java
@@ -1,7 +1,6 @@
 package me.Fupery.ArtMap.Compatibility.impl;
 
 import com.bekvon.bukkit.residence.Residence;
-import com.bekvon.bukkit.residence.containers.Flags;
 import com.bekvon.bukkit.residence.protection.ClaimedResidence;
 import com.bekvon.bukkit.residence.protection.FlagPermissions;
 import com.bekvon.bukkit.residence.protection.ResidencePermissions;